added comments and made small adjustments
