+++ frb_rust/Cargo.toml
@@ -18,7 +18,6 @@ parking_lot = "0.12.1"
 flutter_rust_bridge_macros = { workspace = true }
 uuid = { workspace = true, optional = true }
 log = "0.4"
-type-equals = "0.1.0"
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
 allo-isolate = { workspace = true }
